requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pymongo==4.6.1
transformers==4.38.2
torch==1.13.1
pywhatkit==5.4
python-decouple==3.8
PyPDF2==3.0.1
python-docx==1.0.1
pytesseract==0.3.10
openai
blobfile==2.0.2
sentencepiece==0.1.99
huggingface_hub
python-multipart
twilio
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
bcrypt==4.1.2
python-dotenv
openai
pinecone
pytesseract
pdf2image
pillow
pymupdf